+++ CHANGELOG.rst
@@ -41,7 +41,7 @@ Ver. 0.7.0 - Development
   using python 2.6.
 * Moved the contrib module to :mod:`stdnet.apps`.
 * Added :mod:`stdnet.utils.dates`.
-* **443 regression tests** with **80%** coverage.
+* **477 regression tests** with **80%** coverage.
 
 .. _vers06: